Insert casting operation to be safe.
Modified a bit-shifting operation to cast 1 to a uint64_t.  This
fixed a bug elsewhere in the code; being consistent will hopefully
prevent future bugs.

+++ lib/Transforms/PrivAnalysis/LocalAnalysis.cpp
@@ -49,7 +49,7 @@ void LocalAnalysis::RetrieveAllCAP(CallInst *CI, CAPArray_t &CAPArray)
         unsigned int iarg = I->getZExtValue();
 
         // Add it to the array
-        CAPArray |= 1 << iarg;
+        CAPArray |= ((uint64_t)1u << iarg);
     }
 }
